add styled button component file
