+++ src/components/app/App.js
@@ -1,12 +1,9 @@
 import './App.css';
-import './App.css';
 import {RandomCheese} from '../randomCheese/RandomCheese'
 import LoginForm from '../auth/LoginForm'
 import SignUpForm from '../user/SignUpForm'
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { styled } from '@mui/system';
 import CssBaseline from '@mui/material/CssBaseline';
-import Button from '@mui/material/Button';
 import {
   useNavigate,
   Routes,
@@ -36,15 +33,6 @@ const themeOptions = createTheme({
 //   bg: fg,
 //   });    
 
-// const StyledButton = styled(Button)`
-// color: ${(props) => props.theme.fg};
-// border: 2px solid ${(props) => props.theme.fg};
-// background: ${(props) => props.theme.bg};
-// font-size: 1em;
-// margin: 1em;
-// padding: 0.25em 1em;
-// border-radius: 5px;
-// `;
 
 function App() {
   return (
